update setup-local-dev script

+++ scripts/setup-local-dev.sh
@@ -32,36 +32,76 @@ SOURCE=`pwd`
 
 LOCAL=$SOURCE/local
 APPDIR=$SOURCE/build/webkitbuilds/app.nw
-LOCALCFG=configs/client-default-local.js
 
 
-if [ ! -d $SOURCE/build/webkitbuilds/cache/mac/0.9.3 ]; then
-    mkdir -p $SOURCE/build/webkitbuilds/cache/mac/0.9.3/node-webkit.app
-    pushd $SOURCE/build/webkitbuilds/cache/mac/0.9.3
-    curl -O http://dl.node-webkit.org/v0.9.3/node-webkit-v0.9.3-pre8-osx-ia32.zip
-    unzip node-webkit-v0.9.3-pre8-osx-ia32.zip
+if [ "$os" == "darwin" ]; then
+    if [ ! -d $SOURCE/build/webkitbuilds/cache/mac/0.9.3 ]; then
+        mkdir -p $SOURCE/build/webkitbuilds/cache/mac/0.9.3/node-webkit.app
+        pushd $SOURCE/build/webkitbuilds/cache/mac/0.9.3
+        curl -O http://dl.node-webkit.org/v0.9.3/node-webkit-v0.9.3-pre8-osx-ia32.zip
+        unzip node-webkit-v0.9.3-pre8-osx-ia32.zip
+        popd
+    fi
+    
+    DEST="$SOURCE/build/Cloud9-dev.app"
+    RES="$DEST/Contents/Resources"
+    
+    rm -rf "$DEST"
+    mkdir -p "$RES/app.nw"
+    
+    cp -R $SOURCE/build/webkitbuilds/cache/mac/0.9.3/node-webkit.app/* $DEST
+    cat $SOURCE/local/Info.plist | sed "s/Cloud9/Cloud9-dev/" >  $DEST/Contents/Info.plist
+    # TODO add blue icon for dev mode
+    # rm $DEST/Contents/Resources/nw.icns
+    cp $SOURCE/build/osx/c9.icns $DEST/Contents/Resources/nw.icns
+    
+    node --eval "
+        var path = require('path')
+        var p = require('./local/package.json'); 
+        p.main = path.relative('$RES/app.nw', '$SOURCE/local/projectManager.html');
+        delete p.dependencies;
+        // p.window.icon = 'icon.png';
+        console.log(JSON.stringify(p, null, 2));
+    " > $RES/app.nw/package.json
+    
+    echo dev app created in build/Cloud9-dev.app/Contents/MacOS/node-webkit
+fi
+
+if [ "$os" == "windows" ]; then
+    NODE_VERSION=v0.10.25
+    NW_VERSION=v0.9.2
+    
+    pushd build
+    if [ ! -f node.exe ]; then
+        echo "downloading node"
+        curl -O http://nodejs.org/dist/$NODE_VERSION/node.exe
+    fi
+    if [ ! -f node-webkit-$NW_VERSION-win-ia32.zip ]; then
+        echo "downloading node-webkit"
+        curl -O http://dl.node-webkit.org/$NW_VERSION/node-webkit-$NW_VERSION-win-ia32.zip
+    fi
+    
+    dest=win32-dev/bin
+    mkdir -p $dest
+    
+    unzip node-webkit-$NW_VERSION-win-ia32.zip -d $dest
+    cp node.exe $dest
+    mv $dest/nw.exe $dest/Cloud9.exe
+    
+    cp win32/icon.png $dest
+    
+    # cp -Rf win32/deps win32/bin/deps
+    
+    node --eval '
+        var p=require("../local/package.json"); 
+        p.main = "../../../local/projectManager.html";
+        delete p.dependencies;
+        p.window.icon = "icon.png";
+        console.log(JSON.stringify(p, null, 2));
+    ' > $dest/package.json
+
     popd
 fi
 
-DEST="$SOURCE/build/Cloud9-dev.app"
-RES="$DEST/Contents/Resources"
 
-rm -rf "$DEST"
-mkdir -p "$RES/app.nw"
 
-cp -R $SOURCE/build/webkitbuilds/cache/mac/0.9.3/node-webkit.app/* $DEST
-cat $SOURCE/local/Info.plist | sed "s/Cloud9/Cloud9-dev/" >  $DEST/Contents/Info.plist
-# TODO add blue icon for dev mode
-# rm $DEST/Contents/Resources/nw.icns
-cp $SOURCE/build/osx/c9.icns $DEST/Contents/Resources/nw.icns
-
-node --eval "
-    var path = require('path')
-    var p = require('./local/package.json'); 
-    p.main = path.relative('$RES/app.nw', '$SOURCE/local/projectManager.html');
-    delete p.dependencies;
-    // p.window.icon = 'icon.png';
-    console.log(JSON.stringify(p, null, 2));
-" > $RES/app.nw/package.json
-
-echo dev app created in build/Cloud9-dev.app/Contents/MacOS/node-webkit
